requirements.txt
fastapi
uvicorn[standard]
pymongo
python-dotenv
jinja2
python-multipart
